refactor(webpush): separate delivery contracts from client logic
